+++ packages/all/index.ts
@@ -1,5 +1,9 @@
 const config = {
-	extends: ['@eslint-sets/eslint-config-react', '@eslint-sets/eslint-config-vue']
+	extends: [
+		'@eslint-sets/eslint-config-react',
+		'@eslint-sets/eslint-config-vue',
+		'@eslint-sets/eslint-config-svelte'
+	]
 }
 
 export default config